4th commit: work in progress

+++ get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: cfamilar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/18 12:48:34 by cfamilar          #+#    #+#             */
-/*   Updated: 2022/09/07 18:14:13 by cfamilar         ###   ########.fr       */
+/*   Updated: 2022/09/09 18:43:52 by cfamilar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strchr(char *s, int c)
 {
-	size_t	i;
-	size_t	len;
+	int	i;
+	int	len;
 
 	i = 0;
 	len = ft_strlen(s);
